dom/tests/mochitest/bugs/test_bug1160342_marquee.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1160342
-->
<head>
  <title>Test for Bug 411103</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1160342">Mozilla Bug 1160342</a>
<p id="display"></p>
<div id="content">
<marquee id="a">marquee</marquee>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">
  var x=document.getElementById('a');

  SimpleTest.waitForExplicitFinish();

  setTimeout(function() {
    is(x.behavior, "scroll", "Wrong behavior value");
    x.setAttribute('behavior', 'alternate');
    is(x.behavior, "alternate", "Wrong behavior value");
    x.setAttribute('behavior', 'invalid');
    todo_is(x.behavior, "scroll", "Wrong behavior value");;
    x.setAttribute('behavior', 'Scroll');
    is(x.behavior, "scroll", "Wrong behavior value");
    x.setAttribute('behavior', 'Slide');
    is(x.behavior, "slide", "Wrong behavior value");
    x.removeAttribute('behavior');
    is(x.behavior, "scroll", "Wrong behavior value");

    x.behavior = 'alternate';
    is(x.behavior, "alternate", "Wrong behavior value");
    try {
      x.behavior = 'invalid';
      todo_is(false, true, "marquee.behavior = 'invalid' should throw");
    } catch(e) {
      ok(true, "Exception was raised");
    }
    is(x.behavior, "alternate", "Wrong behavior value");
    x.behavior = 'Slide';
    is(x.behavior, "slide", "Wrong behavior value");
    try {
      x.behavior = 'invalid';
      todo_is(false, true, "marquee.behavior = 'invalid' should throw");
    } catch(e) {
      ok(true, "Exception was raised");
    }
    is(x.behavior, "slide", "Wrong behavior value");

    is(x.loop, -1, "Wrong loop value");
    x.setAttribute('loop', '1');
    is(x.loop, 1, "Wrong loop value");
    x.setAttribute('loop', 'invalid');
    todo_is(x.loop, -1, "Wrong loop value");
    x.setAttribute('loop', '1000');
    is(x.loop, 1000, "Wrong loop value");
    x.setAttribute('loop', '-1');
    is(x.loop, -1, "Wrong loop value");
    x.setAttribute('loop', '1000');
    is(x.loop, 1000, "Wrong loop value");
    x.removeAttribute('loop');
    is(x.loop, -1, "Wrong loop value");

    x.loop = 1;
    is(x.loop, 1, "Wrong loop value");
    try {
      x.loop = -2;
      todo_is(false, true, "marquee.loop = -2 should throw");
    } catch(e) {
      ok(true, "Exception was raised");
    }
    todo_is(x.loop, 1, "Wrong loop value");
    try {
      x.loop = 'invalid';
      todo_is(false, true, ".loop = 'invalid' should throw");
    } catch(e) {
      ok(true, "Exception was raised");
    }
    todo_is(x.loop, 1, "Wrong loop value");
    try {
      x.loop = null;
      todo_is(false, true, "marquee.loop = null should throw");
    } catch(e) {
      ok(true, "Exception was raised");
    }
    todo_is(x.loop, 1, "Wrong loop value");
    x.loop = -1;
    is(x.loop, -1, "Wrong loop value");
    x.loop = '100';
    is(x.loop, 100, "Wrong loop value");

    is(x.scrollAmount, 6, "Wrong scrollAmount value");
    x.setAttribute('scrollAmount', '1');
    is(x.scrollAmount, 1, "Wrong scrollAmount value");
    x.setAttribute('scrollAmount', 'invalid');
    todo_is(x.scrollAmount, 6, "Wrong scrollAmount value");
    x.setAttribute('scrollAmount', '1000');
    is(x.scrollAmount, 1000, "Wrong scrollAmount value");
    x.setAttribute('scrollAmount', '-1');
    is(x.scrollAmount, 1000, "Wrong scrollAmount value");
    x.setAttribute('scrollAmount', '999');
    is(x.scrollAmount, 999, "Wrong scrollAmount value");
    x.removeAttribute('scrollAmount');
    is(x.scrollAmount, 6, "Wrong scrollAmount value");

    x.scrollAmount = 1;
    is(x.scrollAmount, 1, "Wrong scrollAmount value");
    try {
      x.scrollAmount = -2;
      todo_is(false, true, "marquee.scrollAmount = -2 should throw");
    } catch(e) {
      ok(true, "Exception was raised");
    }
    is(x.scrollAmount, 1, "Wrong scrollAmount value");
    x.scrollAmount = 'invalid';
    todo_is(x.scrollAmount, 0, "Wrong scrollAmount value");
    x.scrollAmount = 1;
    x.scrollAmount = null;
    todo_is(x.scrollAmount, 0, "Wrong scrollAmount value");
    x.scrollAmount = '2';
    is(x.scrollAmount, 2, "Wrong scrollAmount value");


    is(x.scrollDelay, 85, "Wrong scrollDelay value");
    x.setAttribute('scrollDelay', '1');
    is(x.scrollDelay, 1, "Wrong scrollDelay value");
    x.setAttribute('scrollDelay', 'invalid');
    todo_is(x.scrollDelay, 85, "Wrong scrollDelay value");
    x.setAttribute('scrollDelay', '70');
    is(x.scrollDelay, 70, "Wrong scrollDelay value");
    x.setAttribute('scrollDelay', '59');
    is(x.scrollDelay, 59, "Wrong scrollDelay value");
    x.setAttribute('scrollDelay', '1000');
    is(x.scrollDelay, 1000, "Wrong scrollDelay value");
    x.setAttribute('scrollDelay', '-1');
    is(x.scrollDelay, 1000, "Wrong scrollDelay value");
    x.removeAttribute('scrollDelay');
    is(x.scrollDelay, 85, "Wrong scrollDelay value");

    x.scrollDelay = 100;
    is(x.scrollDelay, 100, "Wrong scrollDelay value");
    try {
      x.scrollDelay = -2;
      todo_is(false, true, "marquee.scrollDelay = -2 should throw");
    } catch(e) {
      ok(true, "Exception was raised");
    }
    is(x.scrollDelay, 100, "Wrong scrollDelay value");
    try {
      x.scrollDelay = 'invalid';
      todo_is(false, true, "marquee.scrollDelay = 'invalid' should throw");
    } catch(e) {
      ok(true, "Exception was raised");
    }
    is(x.scrollDelay, 100, "Wrong scrollDelay value");
    try {
      x.scrollDelay = null;
      todo_is(false, true, "marquee.scrollDelay = null should throw");
    } catch(e) {
      ok(true, "Exception was raised");
    }
    is(x.scrollDelay, 100, "Wrong scrollDelay value");
    try {
      x.scrollDelay = -1;
      todo_is(false, true, "marquee.scrollDelay = -1 should throw");
    } catch(e) {
      ok(true, "Exception was raised");
    }
    is(x.scrollDelay, 100, "Wrong scrollDelay value");
    x.scrollDelay = '50';
    is(x.scrollDelay, 50, "Wrong scrollDelay value");

    is(x.trueSpeed, false, "Wrong trueSpeed value");
    x.setAttribute('trueSpeed', '1');
    is(x.trueSpeed, true, "Wrong trueSpeed value");
    x.setAttribute('trueSpeed', 'false');
    is(x.trueSpeed, true, "Wrong trueSpeed value");
    x.removeAttribute('trueSpeed');
    is(x.trueSpeed, false, "Wrong trueSpeed value");

    x.trueSpeed = 1;
    is(x.trueSpeed, true, "Wrong trueSpeed value");
    x.trueSpeed = -2;
    is(x.trueSpeed, true, "Wrong trueSpeed value");
    x.trueSpeed = null;
    is(x.trueSpeed, false, "Wrong trueSpeed value");
    x.trueSpeed = '100';
    is(x.trueSpeed, true, "Wrong trueSpeed value");

    todo_is(x.direction, "left", "Wrong direction value");
    x.setAttribute('direction', 'right');
    is(x.direction, "right", "Wrong direction value");
    x.setAttribute('direction', 'invalid');
    todo_is(x.direction, "left", "Wrong direction value");
    x.setAttribute('direction', 'RIGHT');
    todo_is(x.direction, "right", "Wrong direction value");
    x.removeAttribute('direction');
    todo_is(x.direction, "left", "Wrong direction value");

    x.direction = 'right';
    is(x.direction, "right", "Wrong direction value");
    try {
      x.direction = 1;
      todo_is(false, true, "marquee.direction = 1 should throw");
    } catch(e) {
      ok(true, "Exception was raised");
    }
    todo_is(x.direction, "right", "Wrong direction value");
    try {
      x.direction = null;
      todo_is(false, true, "marquee.direction = null should throw");
    } catch(e) {
      ok(true, "Exception was raised");
    }
    todo_is(x.direction, "right", "Wrong direction value");
    SimpleTest.finish();
  }, 0);

</script>
</pre>
</body>
</html>
